totemtv/automation/cloud-run/requirements.txt
flask==3.1.0
gunicorn==23.0.0
google-api-python-client==2.160.0
google-auth==2.37.0
requests==2.32.3
